+++ src/__tests__/__helpers__/fakeClasses.ts
@@ -1,6 +1,10 @@
 // eslint-disable-next-line @typescript-eslint/no-extraneous-class
 export class Bar {}
 
+export class Buzz {
+  public constructor(public name: string) {}
+}
+
 export class Foo {
   public bar: Bar;