+++ src/__tests__/__typetests__/testTypes.test-d.ts
@@ -1,5 +1,6 @@
+// eslint-disable-next-line canonical/filename-match-regex
 import { DIContainer } from '../../DIContainer.js';
-import { Bar } from '../fakeClasses.js';
+import { Bar, Foo } from '../__helpers__/fakeClasses.js';
 import { expectNotType, expectType } from 'tsd';
 import { describe, test } from 'vitest';
 
@@ -27,5 +28,32 @@ describe('DIContainer typescript type resolution', () => {
 
     expectType<Date>(container.a);
     expectNotType<string>(container.a);
+  });
+
+  test('merge containers', () => {
+    const containerA = new DIContainer().add('a', () => 'string');
+    const containerB = new DIContainer().add('b', () => new Date());
+
+    const container = containerA.merge(containerB);
+
+    expectType<Date>(container.b);
+    expectType<string>(container.a);
+  });
+
+  test('extend function', () => {
+    const containerA = () => {
+      return new DIContainer().add('a', () => '1').add('bar', () => new Bar());
+    };
+
+    const finalContainer = containerA().extend((container) => {
+      return container.add('foo', ({ a, bar }) => {
+        return new Foo(a, bar);
+      });
+    });
+
+    // printType(finalContainer);
+    expectType<string>(finalContainer.a);
+    expectType<Bar>(finalContainer.bar);
+    expectType<Foo>(finalContainer.foo);
   });
 });