+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# 3.0.5
+
+Added `merge` and `clone` method to `DIContainer` class.
+
 # 3.0.0
 
 The major release of rsdi version 3.0.0 introduces a revamped API that brings several improvements. The new API aims